Composite recovery tests: range-based SNR + co-locate with other slice/optimizer tests
The composite recovery tests previously used add_intensity_noise (defines
SNR = total_flux / noise_std), while every other intensity slice test uses
add_noise(include_poisson=False) → add_velocity_noise (defines
SNR = (max-min) / noise_std). For typical flux-normalized images this made
the composite test ~100x noisier per pixel at the same nominal SNR=1000.
Fisher information for theta_int dropped from ~4.5e8 (single-comp) to ~60
(composite) at the same labeled SNR — explaining the 3-10% slice biases the
composite test was reporting.

Standardized on the range-based convention via add_noise(include_poisson=False).
Range-based is invariant to image size / pixel scale and is the conventional
"peak per-pixel SNR" definition.

Moved BulgeDisk recovery tests to live alongside other slice/optimizer tests:
- test_likelihood_slice_bulge_disk → test_likelihood_slices.py::test_recover_bulge_disk
- test_optimizer_recovery_bulge_disk → test_optimizer_recovery.py::test_optimize_bulge_disk

test_composite_intensity.py now hosts only composite-specific unit tests +
the constants and synthetic-data helpers (_TRUE_PARS_SHARED, _IMAGE_PARS,
_TEST_PSF, _generate_galsim_composite, _generate_composite_synthetic) that
the recovery tests import.

disk_h_over_r excluded from the slice test (parallels the existing optimizer
exclusion). Justification: at the test geometry projected disk thickness
(sini × h_z ≈ 0.14 arcsec) is comparable to PSF FWHM (0.15 arcsec), so the
parameter is fundamentally near-unobservable post-PSF.

Post-fix recovery: 31/31 unit + 1/1 slice + 2/2 optimizer (incl. SNR=50,
previously catastrophic) all pass at the existing 0.1% / 2% tolerances.
